backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
stripe>=8.0.0
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
pillow==10.1.0
django-filter==23.3
drf-spectacular==0.26.5
gunicorn==21.2.0
whitenoise==6.6.0
django-ratelimit==4.1.0
